fix(FX-127): accept a BLENDINDICES element declared Vector4, not only Byte4
XNA's VertexElementFormat describes the BYTES in the buffer; the shader register a
BLENDINDICES semantic arrives in is a float4 either way, so a content processor is
free to write either. CustomModelAnimation's own SkinnedModelProcessor does
exactly that -- ConvertChannelContent<Vector4>("BlendIndices0") -- and real XNA
renders the result. EasyGL refused the draw:

  EasyGL: this VertexDeclaration cannot be bound to the stock
  'skinned3d_vertexlit' program -- element 1 declares BlendIndices0@12 Vector4 but
  shader input 'aBoneIndices' expects Byte4.

One shader attribute cannot be both an integer and a float, so this is one read
mode on every profile rather than a second program: the three stock skinned vertex
shaders declare "in vec4 aBoneIndices" and cast to int() when indexing uBones[],
Byte4 binds as non-normalized unsigned bytes read as floats (0-255, exactly
representable, far above the <= 72 bone count), and Vector4 binds as plain floats.
That is the shape the GLSL ES 1.00 profiles already ran in production, so the
ES 1.00 rewrite that produced it is gone along with the sources it rewrote.

StockProgramInput grew an alternateFormat, defaulting to format, so a semantic
with more than one legal spelling can say so once.

The CNAEXT engine layer keeps its own copies of the skinned vertex shader in
ShadowMap and DepthNormalPrepass; they are updated with the rest.
ShadowVisibilityTest.ASkinnedMeshShadowsItself is what caught them -- it passes on
the pristine tree, failed with the renderer change alone, and passes again now.

Pinned by easygl_skinnedeffect_vector4_bone_indices_test, whose four checks
separate "it drew" from "it read the index": a Vector4 index of 1 draws and
selects bone 1, index 0 selects the bone that moves the quad off screen, and the
Byte4 spelling still draws. Both halves sabotage-verified -- removing the
acceptance throws again, restoring the integer pointer fails the Byte4 check
alone.

Found by cna-samples SAMPLE-051.

+++ modules/graphics-ext/src/DepthNormalPrepass.cpp
@@ -115,7 +115,10 @@ layout(location = 0) in vec3 aPosition;
 layout(location = 1) in vec3 aNormal;
 layout(location = 2) in vec2 aUV;
 layout(location = 3) in vec4 aBoneWeights;
-layout(location = 4) in uvec4 aBoneIndices;
+// FX-127: a float vec4, matching the stock skinned programs -- XNA's BLENDINDICES is a
+// float4 register whether the declaration spelled the bytes Byte4 or Vector4, so EasyGL
+// binds both as floats and no shader may read the attribute as an integer.
+layout(location = 4) in vec4 aBoneIndices;
 uniform mat4 uWorld;
 uniform mat4 uView;
 uniform mat4 uProjection;
@@ -129,10 +132,10 @@ out float vViewDepth;
 out vec4 vCurrentClip;
 out vec4 vPreviousClip;
 void main() {
-    mat4 skin = uBones[aBoneIndices.x] * aBoneWeights.x;
-    if (uWeightsPerVertex >= 2) skin += uBones[aBoneIndices.y] * aBoneWeights.y;
-    if (uWeightsPerVertex >= 4) skin += uBones[aBoneIndices.z] * aBoneWeights.z
-                                      + uBones[aBoneIndices.w] * aBoneWeights.w;
+    mat4 skin = uBones[int(aBoneIndices.x)] * aBoneWeights.x;
+    if (uWeightsPerVertex >= 2) skin += uBones[int(aBoneIndices.y)] * aBoneWeights.y;
+    if (uWeightsPerVertex >= 4) skin += uBones[int(aBoneIndices.z)] * aBoneWeights.z
+                                      + uBones[int(aBoneIndices.w)] * aBoneWeights.w;
     vec4 world = uWorld * skin * vec4(aPosition, 1.0);
     vec4 view  = uView * world;
     gl_Position = uProjection * view;
